Updated overall ci for new OCI image versioning
Basically it solves how to have oci images which can be tested given a
PR has worked on the ksctl-components directory and also internals on
how can we make it easier to specify from where to get the manifests
instead of the default resolve which is httpGet()

we have a new github action which gets triggered for new PRs provided,
label pr/lgtm is added.

it uses the pr commit head sha and pr number to generate new oci
images which will help in testing in local. will add a
/pr-XYZ as suffix and commit_sha as tag.

modified the oci image builder to utilize the above suffix when it is
recieved via the workflow_dispatch. It also creates a PR comment to
notify the user on how to mmove forward testing the Newly generated oci
artifacts with detailed steps

created new global vars which can be modified via the -ldflags which
will help us to specify that is the commitSha or Release or BranchName
which ever and the suffix which we have constrained to the above.

changed the default value 'latest' ~> 'main' which corresponds to easier
mapping.

also updated the steps for the test/e2e on how to use it

Signed-off-by: Dipankar Das <65275144+dipankardas011@users.noreply.github.com>