Added Poller for fetching releases of github repos
To avoid rate limit of github we added a poller which we can utilize to
get the releases for a given <org>/<repo>, given the order of the
releases to the latest to oldest.

also used singelton reference of the poller interface so that we can
utilize it this object from any location.

Signed-off-by: Dipankar Das <65275144+dipankardas011@users.noreply.github.com>