Added statedocument info about various component versions
Once the object stores the version for example in k3s and kubeadm we are
going to utilize the statedocument as source of truth and it only gets
assigned the object version when creation. Possible future addition is
we need to make sure during the join of worknoders we don't have to
specify the boostrap provider and also the verion it should be able to
get it using the stateDocument. Need to check the manager.go what does the
flow of the invoke of the k8s boostrap looks like.

Signed-off-by: Dipankar Das <65275144+dipankardas011@users.noreply.github.com>