feat(cluster): add cluster summary and direct connect
Introduce `ClusterSummary` in the `common` package to provide a summary
of the cluster, including API server health and node details. Added
`DirectConnect` in the `k8s` package to establish direct connections to
Kubernetes clusters and retrieve health and node summaries.

This change enhances the ability to monitor and manage Kubernetes
clusters effectively.

Signed-off-by: Dipankar Das <65275144+dipankardas011@users.noreply.github.com>